try to integrate markdown editor
