+++ backend/src/resource/page/index.html
@@ -14,9 +14,10 @@
     
     
     
+    
     <link rel="stylesheet" href="/asset/index-ec4ef66164c33b9c.css">
 
-<link rel="preload" href="/asset/index-f136c4aa496c3185_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
-<link rel="modulepreload" href="/asset/index-f136c4aa496c3185.js"></head>
+<link rel="preload" href="/asset/index-b526a6f2685c5051_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
+<link rel="modulepreload" href="/asset/index-b526a6f2685c5051.js"></head>
 <body>
-<script type="module">import init from '/asset/index-f136c4aa496c3185.js';init('/asset/index-f136c4aa496c3185_bg.wasm');</script></body></html>
+<script type="module">import init from '/asset/index-b526a6f2685c5051.js';init('/asset/index-b526a6f2685c5051_bg.wasm');</script></body></html>
