+++ backend/src/service/asset_list.rs
@@ -2,10 +2,11 @@
 ("blog.ico", include_bytes!(r#"..\resource\asset/blog.ico.gz"#), ""),
 ("bulma.min-82aac43507618108.css", include_bytes!(r#"..\resource\asset/bulma.min-82aac43507618108.css.gz"#), "text/css"),
 ("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
+("editor.html", include_bytes!(r#"..\resource\asset/editor.html.gz"#), "text/html; charset=utf-8"),
 ("fontawesome.min-5e9e696c59c57e83.css", include_bytes!(r#"..\resource\asset/fontawesome.min-5e9e696c59c57e83.css.gz"#), "text/css"),
+("index-b526a6f2685c5051.js", include_bytes!(r#"..\resource\asset/index-b526a6f2685c5051.js.gz"#), "text/javascript"),
+("index-b526a6f2685c5051_bg.wasm", include_bytes!(r#"..\resource\asset/index-b526a6f2685c5051_bg.wasm.gz"#), "application/wasm"),
 ("index-ec4ef66164c33b9c.css", include_bytes!(r#"..\resource\asset/index-ec4ef66164c33b9c.css.gz"#), "text/css"),
-("index-f136c4aa496c3185.js", include_bytes!(r#"..\resource\asset/index-f136c4aa496c3185.js.gz"#), "text/javascript"),
-("index-f136c4aa496c3185_bg.wasm", include_bytes!(r#"..\resource\asset/index-f136c4aa496c3185_bg.wasm.gz"#), "application/wasm"),
 ("index.css", include_bytes!(r#"..\resource\asset/index.css.gz"#), "text/css"),
 ("logo.png", include_bytes!(r#"..\resource\asset/logo.png.gz"#), ""),
 ("regular.min-a0c258fb7c5f655d.css", include_bytes!(r#"..\resource\asset/regular.min-a0c258fb7c5f655d.css.gz"#), "text/css"),
