+++ common/src/val.rs
@@ -4,4 +4,4 @@ pub const USER_AUTH_MARK_HEADER: &'static str = "X-SONGDAY-USER-AUTHED";
 pub const POSTS_PAGE_SIZE: u8 = 8;
 pub const DEFAULT_POST_TITLE: &'static str = "未命名/Untitled";
 pub const TAG_SIZES: [&'static str;3] = [" is-normal", " is-medium", " is-large"];
-pub const TAG_COLORS: [&'static str;8] = [" is-dark", " is-light", " is-primary", " is-link", " is-info", " is-success", " is-warning", " is-danger"];
+pub const TAG_COLORS: [&'static str;8] = [" is-white", " is-light", " is-primary", " is-link", " is-info", " is-success", " is-warning", " is-danger"];
