+++ frontend/index.html
@@ -10,6 +10,7 @@
     <link data-trunk rel="css" href="asset/fontawesome.min.css"/>
     <link data-trunk rel="css" href="asset/regular.min.css"/>
     <link data-trunk rel="css" href="asset/solid.min.css"/>
+    <link data-trunk rel="copy-file" href="asset/editor.html"/>
     <link data-trunk rel="copy-file" href="asset/codemirror.min.css"/>
     <link data-trunk rel="copy-file" href="asset/toastui-editor.min.css"/>
     <link data-trunk rel="copy-file" href="asset/toastui-editor-all.min.js"/>
